Cleaning up Profile tests per design guidelines and review feedback
