+++ src/ResourceManager/Profile/Commands.Profile.Test/Commands.Profile.Test.csproj
@@ -192,6 +192,7 @@
     <Compile Include="ErrorResolutionTests.cs" />
     <Compile Include="LocationCompleterUnitTests.cs" />
     <Compile Include="MockDataStore.cs" />
+    <Compile Include="MockServiceClient.cs" />
     <Compile Include="MockSubscriptionClientFactory.cs" />
     <Compile Include="NullClient.cs" />
     <Compile Include="ProfileController.cs" />
